[feat]add ReplyCard
